Create function renderMainPage
